% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylsig_um_predict.R
\name{methylsig_um_predict}
\alias{methylsig_um_predict}
\title{MethylSig-UM Classification Function}
\usage{
methylsig_um_predict(test.data)
}
\arguments{
\item{test.data}{A matrix or data-frame with CpG probe names as named rows and samples as columns.}
}
\description{
Get MethylSig-UM scores (random forest "votes") using MethylSig-UM model
}
\examples{
methylsig_um_predict()
}
\keyword{score}
