% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylsig_um_predict_binary.R
\name{methylsig_um_predict_binary}
\alias{methylsig_um_predict_binary}
\title{MethylSig-UM Binary Classification Function}
\usage{
methylsig_um_predict_binary(test.data)
}
\arguments{
\item{test.data}{A matrix or data-frame with CpG probe names as named rows and samples as columns.}
}
\description{
Get binary classification using MethylSig-UM model
}
\examples{
methylsig_um_predict_binary()
}
\keyword{binary}
